+++ internal/app/wwctl/container/exec/main.go
@@ -124,6 +124,10 @@ func CobraRunE(cmd *cobra.Command, args []string) error {
 		}
 	}
 
+	// getting syncuser from cmd, e.g., shell comamnd will call exec command and passing --syncuser
+	if syncUserFlag, err := cmd.Flags().GetBool("syncuser"); err == nil {
+		SyncUser = SyncUser || syncUserFlag
+	}
 	userdbChanged := false
 	if !beforePasswdTime.IsZero() {
 		afterPasswdTime := getTime(path.Join(containerPath, "/etc/passwd"))
@@ -166,7 +170,7 @@ func getTime(path string) time.Time {
 		return time.Time{}
 	} else {
 		unixStat := fileStat.Sys().(*syscall.Stat_t)
-		return time.Unix(int64(unixStat.Ctim.Sec), int64(unixStat.Ctim.Nsec))
+		return time.Unix(int64(unixStat.Mtim.Sec), int64(unixStat.Mtim.Nsec))
 	}
 }