+++ internal/app/wwctl/container/shell/root.go
@@ -25,6 +25,7 @@ var (
 	}
 	binds    []string
 	nodeName string
+	syncUser bool
 )
 
 func init() {
@@ -33,6 +34,7 @@ Bind a local path which must exist into the container. If destination is not
 set, uses the same path as source. "ro" binds read-only. "copy" temporarily
 copies the file into the container.`)
 	baseCmd.PersistentFlags().StringVarP(&nodeName, "node", "n", "", "Create a read only view of the container for the given node")
+	baseCmd.PersistentFlags().BoolVar(&syncUser, "syncuser", false, "Synchronize UIDs/GIDs from host to container")
 }
 
 // GetRootCommand returns the root cobra.Command for the application.